LV: Hidden methods table
